requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
transformers
sentence-transformers
psycopg2-binary
fastapi
uvicorn[standard]
python-dotenv
pandas
